backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.124.0",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
    "sseclient-py>=1.8.0",
    "uvicorn>=0.38.0",
    "sqlalchemy>=2.0.38",
    "greenlet>=3.3.0",
]
